Moved function to treeview.js
