+++ src/TreeView/treeview.d.ts
@@ -1,0 +1,9 @@
+import { type TreeNode } from "carbon-components-svelte/TreeView/TreeView.svelte";
+/**
+ * Create a nested array from a flat array
+ */
+export function toHierarchy(
+  flatArray: TreeNode[] & { pid?: any }[],
+): TreeNode[];
+
+export default toHierarchy;
